Adding the Lego Import/View/Export capability defined in Sprint 2.
Most of this code is brought directly from the Lego Editor -
https://csfe.aceworkspace.net/svn/repos/lego-editor/trunk/lego-editor

A few small changes were made to line up with the new OTF APIs.  A few
other small GUI tweaks.  No changes to the lego storage / database
model, or anything like that.

There are a few TODOs remanining to be cleaned up, as we integrate
closer with the ISAAC parent app.